pyproject.toml
[project]
name = "mc-webhooks"
version = "0.1.0"
description = "Async webhook listener using FastAPI and Redis"
authors = [{name = "squid1127", email = "him@squid1127.com"}]
readme = "README.md"

requires-python = ">=3.10,<4.0"

dependencies = [
    "redis>=5.0.0",
    "pydantic>=1.10.7",
    "typer (>=0.21.1)",
    "pydantic-settings (>=2.12.0)",
    "fastapi (>=0.128.0)",
    "uvicorn[standard] (>=0.40.0)",
    "discord-py (>=2.6.4,<3.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[tool.poetry.scripts]
mc_webhooks = "mc_webhooks.main:app"
mcwh = "mc_webhooks.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
